feat(*): drop dialog / popup support, fix almost all linter warnings, more general cleanup
BREAKING CHANGE: This commit drops support for opening the DateTimePicker in a Popup/Dialog.
If you need this functionality, just wrap the inline picker in a CDK-Overlay or CDK-Dialog yourself.
Next up, the whole "UnixTimestampDateTimeAdapter" (+ related classes) were deprecated and will be removed soon
(Just use the NativeDateTimeAdapter instead). The `localeChanges` getter that returned an observable which emits
whenever the locale changes was refactored to a property and renamed to `localeChanges$` to indicate that it's an observable.
The `owlDateTimeFilter` was renamed to `dateFilter` because it was the only input that was prefixed with `owl`.
Last but not least, it fixes many internal linter warnings and adds proper typing to the project. Many (not all 🥲) `any`
declarations were replaced with proper types which can be considered a breaking change aswell.