feat(*): update Angular, Drop dialog support, move from scss to css
BREAKING CHANGE: Clients need to update to Angular 20 in order to use this version. As I wrote last october in #3, Dialog and Popup support were finally dropped. Only the "inline" mode is supported from now on. Additionally, probably irrelevant for clients, this library switched from scss to css which simplifies the build process.